Move "Optional" to front
This is purely my personal opinion and meant as a suggestion rather than a fix!
I would want the "Optional" in front so when I just want to edit one line in the docs, I know right away that I can skip it.

Again: Personal opinion. Feel free to drop.