add a note that this is not docker but npm
Being new to the project, I assumed this meant I could run my Docker builds there as well. The reason I needed this is because when I install Supabase, it can really lock up the server. So I was hoping this could be a way to do it on another server, but then use it on the server I'm on. But I realized now I don't need to make it a build server. I could just use it as a normal other server.