comment out httpchallenge in traefik config example
My setup didnt work until I made sure to only use one challenge for the letsencrypt resolver. Internet seems to point in the direction  that traefik isnt happy if setting both a dns and http challenge for a resolver. I havent verified this in traefik dos, but it started working for me when making the change.